requirements.txt
opencv-python
torch==2.2.1
fastapi==0.110.0
Pillow==10.2.0
pandas==2.0.3
"uvicorn[standard]"
opencv-python==
scikit-learn
python-multipart